third_party/blink/web_tests/platform/generic/storage/indexeddb/deleteObjectStore-name-argument-required-expected.txt

Test IndexedDB deleteObjectStore required argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "deleteObjectStore-name-argument-required.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
PASS db.deleteObjectStore(); threw exception TypeError: Failed to execute 'deleteObjectStore' on 'IDBDatabase': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
